# public/openapi.yaml
openapi: 3.0.3
info:
  title: ConejoMalo API
  version: '1.0.0'
  description: |
    API pública y de administración para la tienda. Autenticación vía cookie `session`.
    Este documento es la fuente de verdad de los payloads.
servers:
  - url: https://{host}/api
    variables:
      host:
        default: localhost:3000
  - url: http://localhost:3000/api
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string, nullable: true }
        fields:
          type: object
          additionalProperties: { type: string }
          nullable: true
...
    UserListItem:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        roles:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        price: { oneOf: [ { type: string }, { type: number } ] }
        imageUrl: { type: string, nullable: true }
        categoryId: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        category:
          type: object
          nullable: true
          properties:
            id: { type: string }
            name: { type: string }
            slug: { type: string }
...
paths:
  /products:
    get:
      summary: Listar productos públicos
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per
          schema: { type: integer, minimum: 1, maximum: 60 }
        - in: query
          name: sort
          schema: { type: string, enum: [new, price_asc, price_desc, name] }
        - in: query
          name: q
          schema: { type: string, minLength: 2 }
        - in: query
          name: cat
          schema: { type: string }
      responses:
        '200':
          description: Lista o paginación
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/Product' }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Product' }
                      page: { type: integer }
                      per: { type: integer }
                      total: { type: integer }
                      totalPages: { type: integer }
        '500': { description: Error, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
  /admin/users:
    get:
      summary: Listar usuarios (admin)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/UserListItem' } }, page: { type: integer }, per: { type: integer }, total: { type: integer }, totalPages: { type: integer } } } } } }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
    post:
      summary: Crear usuario (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                name: { type: string, nullable: true }
                roles:
                  type: array
                  items: { type: string }
                password: { type: string }
      responses:
        '201': { description: Creado, content: { application/json: { schema: { $ref: '#/components/schemas/UserListItem' } } } }
        '400': { description: Error de validación }
        '409': { description: Ya existe }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
...
security:
  - cookieAuth: []